# contracts/nexus-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Nexus AI Agent Office API",
    "description": "Backend API for the AI Agent Office OS SaaS Platform",
    "version": "v1"
  },
  "paths": {
...
    "/api/Agents/office/{officeId}": {
      "get": {
        "tags": [
          "Agents"
        ],
        "parameters": [
          {
            "name": "officeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Office/{id}": {
      "get": {
        "tags": [
          "Office"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OfficeDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfficeDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfficeDetailDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Tasks": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 80
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Tasks/brief/{briefId}": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "briefId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Tasks/{id}": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Tasks/{id}/status": {
      "put": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskStatusRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskStatusRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Tasks/{id}/assign": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignTaskRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignTaskRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AssignTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AgentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          },
          "agentType": {
            "$ref": "#/components/schemas/AgentType"
          },
          "state": {
            "$ref": "#/components/schemas/AgentState"
          },
          "isEnabled": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "AgentState": {
        "enum": [
          "Idle",
          "Working",
          "Collaborating",
          "Blocked",
          "Completed",
          "Error",
          "Offline"
        ],
        "type": "string"
      },
      "AgentType": {
        "enum": [
          "AiCeo",
          "BlogWriter",
          "SocialMediaDesigner",
          "InstagramContentGenerator",
          "UiUxDesigner",
          "VideoEditor",
          "SeoSpecialist",
          "GoogleAdsAnalyst",
          "CustomerReviewResponder",
          "ChatbotManager",
          "AiStrategist",
          "AnalyticsAnalyst",
          "ContentStrategy"
        ],
        "type": "string"
      },
...
      "AssignTaskRequest": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string",
            "format": "uuid"
          },
          "agentId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "OfficeDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isDefault": {
            "type": "boolean"
          },
          "zones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OfficeZoneDto"
            }
          },
          "agents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentDto"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "OfficeZoneDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "zoneType": {
            "$ref": "#/components/schemas/OfficeZoneType"
          },
          "name": {
            "type": "string"
          },
          "positionX": {
            "type": "number",
            "format": "double"
          },
          "positionY": {
            "type": "number",
            "format": "double"
          },
          "positionZ": {
            "type": "number",
            "format": "double"
          },
          "width": {
            "type": "number",
            "format": "double"
          },
          "depth": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "OfficeZoneType": {
        "enum": [
          "CommandCenter",
          "ContentStudio",
          "DesignLab",
          "MediaBay",
          "AnalyticsFloor",
          "CommunicationHub"
        ],
        "type": "string"
      },
...
      "TaskDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "briefId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "agentType": {
            "$ref": "#/components/schemas/AgentType"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "priority": {
            "$ref": "#/components/schemas/TaskPriority"
          },
          "estimatedDurationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "actualDurationMinutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "errorMessage": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TaskPriority": {
        "enum": [
          "Low",
          "Normal",
          "High",
          "Urgent",
          "Critical"
        ],
        "type": "string"
      },
      "TaskStatus": {
        "enum": [
          "Pending",
          "Queued",
          "InProgress",
          "WaitingForDependency",
          "WaitingForApproval",
          "Approved",
          "Rejected",
          "RevisionRequested",
          "Completed",
          "Failed",
          "Cancelled"
        ],
        "type": "string"
      },
...
      "UpdateTaskStatusRequest": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}